+++ app/webpacker/components/EditSchedule/ManageActivities/ActionsHeader.js
