+++ app/webpacker/components/EditSchedule/ManageActivities/ActivityPicker.js
