+++ app/webpacker/components/EditSchedule/ManageActivities/EditActivityModal.js
