+++ app/webpacker/components/EditSchedule/ManageActivities/index.js
